% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_tests.R
\name{chisq_test}
\alias{chisq_test}
\alias{chisq_test.default}
\alias{chisq_test.matrix}
\title{Chisq functions}
\usage{
chisq_test(observed, ...)

\method{chisq_test}{default}(observed, expected, ...)

\method{chisq_test}{matrix}(observed, ...)
}
\arguments{
\item{observed}{This is the observed counts}

\item{...}{Any other varaibles. For example \code{expected} which is used in the default method.}

\item{expected}{This is the expected counts}
}
\value{
in almost all cases it returns the p-value
}
\description{
Various homebrew functions to work with chisq tests
}
\details{
the p-value from a chisq test for the observed and expected outcomes.

Get the p-value from the contingency table for whether the rows and colums are associated or not.
}
