% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_t_dist.R
\name{CI}
\alias{CI}
\alias{CI_t_dist}
\alias{print.CI_t_dist}
\title{Quick function to get population mean estimate from sample}
\usage{
CI_t_dist(data, level = 0.95)

\method{print}{CI_t_dist}(x, ...)
}
\arguments{
\item{data}{A atomic vector}

\item{level}{Confidence level you would like. Standard is 0.95 (95\%) however it is a aribatary number.}

\item{x}{CI_t_dist object to be printed}

\item{...}{Ignored by this printing function}
}
\value{
CI_t_dist class object with all the required information

Invisible CI_t_dist object.
}
\description{
This function will get the confidence interval for a sample data set.
}
\examples{
exampleData <- rnorm(1e4)
ci <- CI_t_dist(exampleData)
print(ci)
}
