% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspondence_analysis.R
\name{correspondence_analysis}
\alias{correspondence_analysis}
\alias{data_profile}
\alias{plot.table_profile}
\title{Get the profiles of a 2d table}
\usage{
data_profile(data, rowprofile = TRUE)

\method{plot}{table_profile}(x, y, ...)
}
\arguments{
\item{data}{The 2d table you want to look at}

\item{rowprofile}{Whether you want it to be a row or col profile. False is for column profile}

\item{x}{The 2d profile table}

\item{y}{This is ignored by this plot function}

\item{...}{This is also ignored by this plot function}
}
\value{
A 2d matrix with the class \code{table_profile}

Plot returns nothing
}
\description{
Used to get either the row or column profiles of a data table.
}
\details{
It currently it designed to work with matrixes. There ought to be row names and col names for it to plot correctly.

The plot function will simply plot this in a way that you would expect
}
