% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_stats.R
\name{D_stats}
\alias{D_stats}
\title{Find out the three D-stats for symmetry}
\usage{
D_stats(x)
}
\arguments{
\item{x}{The variable that you are going to be looking at.}
}
\value{
a named vector with the three d values
}
\description{
This gets a vector with the three d values.
}
\details{
The three different d values are useful for different purposes.
D1 is not very robust but quite simple to understand and comes from \code{(mean(x) - median(x))/sd(x)}.
D2 is more robust by instead using the f-spread \code{(mean(x) - median(x))/(fivenum(x)[4] - fivenum(x)[2])}.
D3 lastly only measures the the middle 50\% of the data. \code{((fivenum(x)[4]+fivenum(x)[2])/2 - median(x))/(fivenum(x)[4] - fivenum(x)[2])}
}
\examples{
# Find out if the height is symmetric
D_stats(Loblolly$height)

}
