% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_tests.R
\name{.format_chisq_output}
\alias{.format_chisq_output}
\title{Structure the chisq_output}
\usage{
.format_chisq_output(chisq_value = NULL, df = NULL, p_value = NULL)
}
\arguments{
\item{chisq_value}{The chisq value that you have gotten from the calculations}

\item{df}{the degrees of freedom}

\item{p_value}{the p_value after applying to the the pchisq function. This can be left null if it is than it will be calculated.}
}
\value{
A list structure with the class \code{chisq_output}
}
\description{
Makes the formatting of the chisq outputs central so that all the smae names can be used without worry.
}
