% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_workflow.r
\name{.renderAndStyle}
\alias{.renderAndStyle}
\title{Function to prepare the pkg for a commit}
\usage{
.renderAndStyle(location = ".")
}
\arguments{
\item{location}{Where the base directory is for the package you want to prepare}
}
\description{
This will style the code and render the readme and helps.
}
