% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_tests.R
\name{print.chisq_output}
\alias{print.chisq_output}
\title{Printout for the chisq_output}
\usage{
\method{print}{chisq_output}(x, ...)
}
\arguments{
\item{x}{A chisq_output that you want to print out}

\item{...}{Any other variable currently they are ignored.}
}
\value{
the \verb{invisible object}
}
\description{
Printout for the chisq_output
}
