% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals}
\alias{residuals}
\alias{residuals_list}
\alias{plot.residuals.data}
\alias{print.residuals.data}
\title{Create residual table}
\usage{
residuals_list(lm = NULL)

\method{plot}{residuals.data}(x, ...)

\method{print}{residuals.data}(x, ...)
}
\arguments{
\item{lm}{A linear model object created by the \code{lm()} function}

\item{x}{A \code{residuals.data} list object}

\item{...}{ignored by this method}
}
\value{
A residuals

Plot function will return the plot object

A plain data.frame
}
\description{
Theses functions are created to make the calculating and displaying of
residuals easy and quick. It will get the raw, standard and studentized
residuals from the lm object.
}
\examples{
mtcars.mod <- lm(mpg ~ wt, data = mtcars)
residuals <- residuals_list(mtcars.mod)
plot(residuals)
print(residuals)
}
