% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.161250-package.R
\docType{package}
\name{util.161250-package}
\alias{util.161250}
\alias{util.161250-package}
\title{util.161250: Producitivity Functions for 161250}
\description{
Functions that have been made to speed up my course work for 161.250 at Massey University.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/1jamesthompson1/util.161250}
  \item Report bugs at \url{https://github.com/1jamesthompson1/util.161250/issues}
}

}
\author{
\strong{Maintainer}: James Thompson \email{1jamesthompson1@gmail.com} (\href{https://orcid.org/0000-0002-4768-795X}{ORCID})

}
\keyword{internal}
